+++ src/main/java/online/kingdomkeys/kingdomkeys/client/gui/overlay/CommandMenuGui.java
@@ -736,6 +736,7 @@ public class CommandMenuGui extends OverlayBase {
 		ResourceLocation rcTexture = commandMenuElements.get(currentSubmenu).getTexture();
 
 		int i = 0;
+		boolean valid = false;
 		for (Map.Entry<String, Integer> entry : list.entrySet()) {
 			gui.pose().pushPose();
 			{
@@ -747,27 +748,39 @@ public class CommandMenuGui extends OverlayBase {
 				{
 					ReactionCommand command = ModReactionCommands.registry.get(ResourceLocation.parse(entry.getKey()));
 					String time = "";
-					if(entry.getValue() > -1){
+					if (entry.getValue() > -1) {
 						time = String.format("%.1f", entry.getValue() / 20.0);
 						gui.pose().pushPose();
 						gui.pose().scale(0.6F, 0.8F, scale);
-						drawString(gui, minecraft.font, Component.literal(time).withStyle(ClientUtils.KK_Font_EXP),(int)((TOP_WIDTH - ModConfigs.cmReactionEndRWidth) * 1.9F), 6, 0xFFFFFF);
+						drawString(gui, minecraft.font, Component.literal(time).withStyle(ClientUtils.KK_Font_EXP), (int) ((TOP_WIDTH - ModConfigs.cmReactionEndRWidth) * 1.9F), 6, 0xFFFFFF);
 						gui.pose().popPose();
 					}
-					drawString(gui, minecraft.font, Utils.translateToLocal(command.getTranslationKey()),5 + (ModConfigs.cmTextXOffset+5), 4, 0xFFFFFF);
+					drawString(gui, minecraft.font, Utils.translateToLocal(command.getTranslationKey()), ModConfigs.cmTextXOffset + 15, 4, 0xFFFFFF);
 
 					gui.pose().scale(1.33F, 1, 1);
 					RenderSystem.enableBlend();
-					blit(gui, rcTexture, 0, 0, 0, 45, ModConfigs.cmReactionEndLWidth+1, TOP_HEIGHT);
+					blit(gui, rcTexture, 0, 0, 0, 45, ModConfigs.cmReactionEndLWidth + 1, TOP_HEIGHT);
 					int middleWidth = TOP_WIDTH - (ModConfigs.cmReactionEndLWidth + ModConfigs.cmReactionEndRWidth) + 1;
-					if(time.isEmpty()) {
+					if (time.isEmpty()) {
 						blit(gui, rcTexture, ModConfigs.cmReactionEndLWidth, 0, middleWidth, TOP_HEIGHT, ModConfigs.cmReactionEndLWidth + 1, 45, 1, TOP_HEIGHT, 256, 256);
 					} else {
-						blit(gui, rcTexture, ModConfigs.cmReactionEndLWidth, 0, middleWidth, TOP_HEIGHT, 35, 45, 1, TOP_HEIGHT, 256, 256);
+						blit(gui, rcTexture, ModConfigs.cmReactionEndLWidth, 0, middleWidth, TOP_HEIGHT, 24, 45, 1, TOP_HEIGHT, 256, 256);
 						float perc = 100F * entry.getValue() / command.getDuration();
-						blit(gui, rcTexture, ModConfigs.cmReactionEndLWidth, 0, (int)(middleWidth * perc/100F), TOP_HEIGHT, 37, 45, 1, TOP_HEIGHT, 256, 256);
+						blit(gui, rcTexture, ModConfigs.cmReactionEndLWidth, 0, (int) (middleWidth * perc / 100F), TOP_HEIGHT, 26, 45, 1, TOP_HEIGHT, 256, 256);
 					}
 					blit(gui, rcTexture, TOP_WIDTH - ModConfigs.cmReactionEndRWidth, 0, ModConfigs.cmReactionEndLWidth + 3, 45, ModConfigs.cmReactionEndRWidth, TOP_HEIGHT);
+
+					if(i == reactionSelected) {
+						valid = true;
+						gui.pose().pushPose();
+						{
+							gui.pose().scale(0.7F, 1F, 1);
+							blit(gui, rcTexture, 4, 3, 28, 45, 10, 10);
+							gui.drawString(minecraft.font, Component.literal(InputHandler.Keybinds.REACTION_COMMAND.getKeybind().getKey().getDisplayName().getString()).withStyle(ClientUtils.KK_Font_MENU), 6, 3, 0xFFFFFF, false);
+						}
+						gui.pose().popPose();
+					}
+
 					RenderSystem.disableBlend();
 				}
 				gui.pose().popPose();
@@ -775,6 +788,9 @@ public class CommandMenuGui extends OverlayBase {
 			}
 			gui.pose().popPose();
 			i++;
+		}
+		if(!valid) {
+			reactionSelected = 0;
 		}
 
 	}