+++ src/main/java/online/kingdomkeys/kingdomkeys/entity/magic/WaterzaEntity.java
@@ -149,7 +149,7 @@ public class WaterzaEntity extends ThrowableProjectile {
 						}
 						if(p == null || (p.getMember(target.getUUID()) == null || p.getFriendlyFire())) { //If caster is not in a party || the party doesn't have the target in it || the party has FF on
 							float dmg = this.getOwner() instanceof Player ? DamageCalculation.getMagicDamage((Player) this.getOwner()) * 1.3F : 2;
-							target.hurt(target.damageSources().thrown(this, this.getOwner()), dmg * dmgMult);
+							target.hurt(KKDamageTypes.getElementalDamage(KKDamageTypes.WATER,this, this.getOwner()), dmg * dmgMult);
 						}
 					}
 				}
@@ -202,7 +202,7 @@ public class WaterzaEntity extends ThrowableProjectile {
 							if(!Utils.isEntityInParty(casterParty, e) && e != player) {
 								float baseDmg = DamageCalculation.getMagicDamage((Player) this.getOwner()) * 1.2F;
 								float dmg = this.getOwner() instanceof Player ? baseDmg : 2;
-								e.hurt(e.damageSources().thrown(this, this.getOwner()), dmg * dmgMult);
+								e.hurt(KKDamageTypes.getElementalDamage(KKDamageTypes.WATER,this, this.getOwner()), dmg * dmgMult);
 							}
 						}
 					}